backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.54
pydantic==1.10.13
python-dotenv==1.0.1
aiofiles==23.1.0

